+++ bin/test.sh
@@ -14,3 +14,14 @@ for script in test-scripts/*.lua; do
 	echo "Running $(basename "$script" .txt)"
 	./target/debug/remodel "$script"
 done
+
+if [ ! -z "${REMODEL_AUTH_TESTS}" ]; then
+	echo ""
+	echo "Running extra tests that need network access"
+	echo ""
+
+	for script in test-scripts-extra/*.lua; do
+		echo "Running $(basename "$script" .txt)"
+		./target/debug/remodel "$script"
+	done
+fi