+++ test-scripts-extra/write-asset-model.lua
@@ -1,0 +1,3 @@
+local source = remodel.readModelFile("test-models/folder-and-value.rbxmx")[1]
+
+remodel.writeExistingModelAsset(source, "3762610069")