+++ test-scripts-extra/write-asset-place.lua
@@ -1,0 +1,3 @@
+local source = remodel.readPlaceFile("test-models/place-with-models.rbxlx")
+
+remodel.writeExistingPlaceAsset(source, "2983428072")