+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Remodel Changelog
 
 ## Unreleased Changes
+* Added APIs for interacting with models and places on Roblox.com:
+	* `remodel.readModelAsset`
+	* `remodel.writeExistingModelAsset`
+	* `remodel.writeExistingPlaceAsset`
 
 ## 0.4.0 (2019-09-18)
 * Added `remodel.readDir` for enumerating directories.